test: simplify screenshot attachment helper
